+++ .gitignore
@@ -2,3 +2,4 @@
 target
 *.iml
 Cargo.lock
+client.db.*
